+++ data/collisions.js
@@ -5,28 +5,28 @@ export const   floorCollisions   =   [
             0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  , 
             0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  , 
             0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  , 
+            0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,   202 ,   202 ,    0  ,    0  ,    0  , 
+            0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  , 
+            0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  , 
             0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  , 
+            0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  , 
+            0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,   202 ,   202 , 
+            0  ,    0  ,    0  ,    0  ,    0  ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,   202 ,   202 ,   202 ,   202 , 
+            0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,   202 ,   202 ,   202 ,   202 ,   202 ,   202 ,   202 ,   202 ,   202 ,   202 ,   202 ,   202 ,   202 ,   202 , 
             0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  , 
-            0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  , 
-            0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  , 
-            0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  , 
-            0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  , 
-            0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  , 
-            0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  , 
-            0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  , 
-            0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,
-            0  ,    0  ,    0  ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,
-            0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,
-            0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,
+            0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,
+            0  ,    0  ,    0  ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,
             0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,
             0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,
-            0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,
+            0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,
             0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,
             0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,
             0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,
-            0  ,    0  ,    0  ,    0  ,    0  ,    0  ,   202 ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,
-            0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,   202 ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,
+            0  ,    0  ,    0  ,   202 ,   202 ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,   202 ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,
+            0  ,    0  ,    0  ,    0  ,    0  ,    0  ,   202 ,    0  ,    0  ,    0  ,    0  ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,
+            0  ,    0  ,    0  ,    0  ,    0  ,    0  ,   202 ,   202 ,   202 ,   202 ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,
             0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,   202 ,   202 ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,
-            0  ,    0  ,    0  ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,   202 ,   202 ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,
-            0  ,    0  ,    0  ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,   202 ,   202 ,   202 ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,
+           202 ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,   202 ,   202 ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,
+           202 ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,   202 ,   202 ,   202 ,   202 ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,    0  ,
+           202 ,   202 ,   202 ,   202 ,   202 ,   202 ,   202 ,   202 ,   202 ,   202 ,   202 ,   202 ,   202 ,    0  ,    0  ,   202 ,   202 ,   202 ,   202 ,   202 ,   202 ,   202 ,   202 ,   202 ,    0  ,    0  ,   202 ,   202 ,   202 ,   202 ,
 ]
