+++ js/collisionBlocks.js
@@ -1,5 +1,5 @@
 export class CollisionBlock {
-  constructor({ x, y, ctx, width = 36, height = 36 }) {
+  constructor({ x, y, ctx, width, height}) {
     this.x = x
     this.y = y
     this.width = width
@@ -7,7 +7,7 @@ export class CollisionBlock {
     this.context = ctx
   }
   draw() {
-    this.context.fillStyle = 'blue'
+    this.context.fillStyle = 'grey'
     this.context.fillRect(this.x, this.y, this.width, this.height)
   }  
   update() {
